feat: refactor transformPropertyToRules
